test_requirements.txt
Flask-Testing==0.4.1
MiniMock==1.2.8
coverage==3.7.1
fakeredis==0.4.2
nose==1.3.0
